Change templated related_name fields in model migrations
Abstract class models require that the `related_name` arg for ForeignKey fields be configured with a templated string. In Django 3.2.x (used by Nautobot up through version 2.2.x), `makemigration` will replace that with the actual string name in the migration file. In Django 4 (used by Nautobot 2.3.x), the migration state check has been overhauled, and it no longer recognizes the replacement string as the correct value, wanting you to run `makemigrations` again to add a new migration file which sets the option back to the template string instead. The migration itself won’t make any changes, as when the migration runs it will replace the template string with the actual string.

This update sets the initial migration file to have the template string, so that when upgrading to Nautobot 2.3, Django will no longer have spurious complaints about the migrations.

Signed-off-by: Mark Troyer <mtroyer@nvidia.com>